+++ backend/canisters/local_user_index/impl/src/updates/join_channel.rs
@@ -1,15 +1,13 @@
 use crate::guards::caller_is_openchat_user;
-use crate::{mutate_state, read_state, RuntimeState};
-use candid::Principal;
+use crate::{mutate_state, read_state};
 use canister_tracing_macros::trace;
 use ic_cdk_macros::update;
 use local_user_index_canister::join_channel::{Response::*, *};
-use types::UserId;
 
 #[update(guard = "caller_is_openchat_user")]
 #[trace]
 async fn join_channel(args: Args) -> Response {
-    let user_details = read_state(user_details);
+    let user_details = read_state(|state| state.calling_user());
 
     let c2c_args = community_canister::c2c_join_channel::Args {
         user_id: user_details.user_id,
@@ -17,7 +15,7 @@ async fn join_channel(args: Args) -> Response {
         channel_id: args.channel_id,
         invite_code: args.invite_code,
         is_platform_moderator: user_details.is_platform_moderator,
-        rules_accepted: args.rules_accepted,
+        is_bot: user_details.is_bot,
     };
     match community_canister_c2c_client::c2c_join_channel(args.community_id.into(), &c2c_args).await {
         Ok(response) => match response {
@@ -38,25 +36,8 @@ async fn join_channel(args: Args) -> Response {
             community_canister::c2c_join_channel::Response::MemberLimitReached(l) => MemberLimitReached(l),
             community_canister::c2c_join_channel::Response::CommunityFrozen => CommunityFrozen,
             community_canister::c2c_join_channel::Response::NotInvited => NotInvited,
-            community_canister::c2c_join_channel::Response::RulesNotAccepted => RulesNotAccepted,
             community_canister::c2c_join_channel::Response::InternalError(error) => InternalError(error),
         },
         Err(error) => InternalError(format!("Failed to call 'community::c2c_join_channel': {error:?}")),
-    }
-}
-
-struct UserDetails {
-    user_id: UserId,
-    principal: Principal,
-    is_platform_moderator: bool,
-}
-
-fn user_details(state: &RuntimeState) -> UserDetails {
-    let user = state.calling_user();
-
-    UserDetails {
-        user_id: user.user_id,
-        principal: user.principal,
-        is_platform_moderator: user.is_platform_moderator,
     }
 }