+++ backend/integration_tests/src/communities/send_message_tests.rs
@@ -7,7 +7,8 @@ use ic_test_state_machine_client::StateMachine;
 use ledger_utils::create_pending_transaction;
 use std::ops::Deref;
 use types::{
-    ChannelId, ChatEvent, CommunityId, CryptoContent, CryptoTransaction, Cryptocurrency, MessageContent, MessageContentInitial,
+    AccessRules, ChannelId, ChatEvent, CommunityId, CryptoContent, CryptoTransaction, Cryptocurrency, MessageContent,
+    MessageContentInitial, OptionUpdate, TextContent, Version,
 };
 
 #[test]
@@ -83,7 +84,8 @@ fn send_crypto_in_channel() {
             sender_name: user1.username(),
             replies_to: None,
             mentioned: Vec::new(),
-            rules_accepted: None,
+            community_rules_accepted: None,
+            channel_rules_accepted: None,
         },
     );
 
@@ -96,6 +98,459 @@ fn send_crypto_in_channel() {
     } else {
         panic!("{send_message_result:?}")
     }
+}
+
+#[test]
+fn send_message_with_community_rules_not_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+
+    let _response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, None);
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_channel_rules_not_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    let _response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, None);
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_community_rules_accepted_succeeds() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+
+    let response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, Some(Version::from(1)), None);
+
+    if !matches!(response, community_canister::send_message::Response::Success(_)) {
+        panic!("'send_message' error {response:?}");
+    }
+}
+
+#[test]
+fn send_message_with_channel_rules_accepted_succeeds() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    let response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, Some(Version::from(1)));
+
+    if !matches!(response, community_canister::send_message::Response::Success(_)) {
+        panic!("'send_message' error {response:?}");
+    }
+}
+
+#[test]
+fn send_message_with_community_rules_but_not_channel_rules_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    let _response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, Some(Version::from(1)), None);
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_channel_rules_but_not_community_rules_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    let _response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, Some(Version::from(1)));
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_community_rules_and_channel_rules_accepted_succeeds() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    let response = send_dummy_message_with_rules(
+        env,
+        &user2,
+        community_id,
+        channel_id,
+        Some(Version::from(1)),
+        Some(Version::from(1)),
+    );
+
+    if !matches!(response, community_canister::send_message::Response::Success(_)) {
+        panic!("'send_message' error {response:?}");
+    }
+}
+
+#[test]
+fn send_message_with_previously_accepted_rules_succeeds() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+
+    send_dummy_message_with_rules(
+        env,
+        &user2,
+        community_id,
+        channel_id,
+        Some(Version::from(1)),
+        Some(Version::from(1)),
+    );
+
+    let response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, None);
+
+    if !matches!(response, community_canister::send_message::Response::Success(_)) {
+        panic!("'send_message' error {response:?}");
+    }
+}
+
+#[test]
+fn send_message_with_old_community_rules_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_community_rules(env, &user1, community_id, "No heavy petting".to_string());
+    set_community_rules(env, &user1, community_id, "No heavy petting or pets".to_string());
+
+    let _response = send_dummy_message_with_rules(
+        env,
+        &user2,
+        community_id,
+        channel_id,
+        Some(Version::from(1)),
+        Some(Version::from(1)),
+    );
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_old_channel_rules_accepted_fails() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    set_channel_rules(env, &user1, community_id, channel_id, "No running".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No running or jumping".to_string());
+
+    let _response = send_dummy_message_with_rules(env, &user2, community_id, channel_id, None, Some(Version::from(1)));
+
+    // TODO: Re-enable check once RulesNotAccepted is returned
+    // if !matches!(response, community_canister::send_message::Response::RulesNotAccepted) {
+    //     panic!("{response:?}");
+    // }
+}
+
+#[test]
+fn send_message_with_rules_leads_to_expected_summary_and_selected_states() {
+    let mut wrapper = ENV.deref().get();
+
+    let TestEnv {
+        env,
+        canister_ids,
+        controller,
+    } = wrapper.env();
+
+    let TestData {
+        user1,
+        user2,
+        community_id,
+        channel_id,
+    } = init_test_data(env, canister_ids, *controller);
+
+    let community_rules = get_community_rules(env, &user2, community_id);
+    let channel_rules = get_channel_rules(env, &user2, community_id, channel_id);
+
+    assert_eq!(
+        community_rules,
+        Rules {
+            enabled: false,
+            accepted: false,
+            text: "".to_string(),
+            version: Version::zero()
+        }
+    );
+
+    assert_eq!(
+        channel_rules,
+        Rules {
+            enabled: false,
+            accepted: false,
+            text: "".to_string(),
+            version: Version::zero()
+        }
+    );
+
+    set_community_rules(env, &user1, community_id, "No running".to_string());
+    set_channel_rules(env, &user1, community_id, channel_id, "No jumping".to_string());
+
+    let community_rules = get_community_rules(env, &user2, community_id);
+    let channel_rules = get_channel_rules(env, &user2, community_id, channel_id);
+
+    assert_eq!(
+        community_rules,
+        Rules {
+            enabled: true,
+            accepted: false,
+            text: "No running".to_string(),
+            version: Version::from(1)
+        }
+    );
+
+    assert_eq!(
+        channel_rules,
+        Rules {
+            enabled: true,
+            accepted: false,
+            text: "No jumping".to_string(),
+            version: Version::from(1)
+        }
+    );
+
+    send_dummy_message_with_rules(
+        env,
+        &user2,
+        community_id,
+        channel_id,
+        Some(Version::from(1)),
+        Some(Version::from(1)),
+    );
+
+    let community_rules = get_community_rules(env, &user2, community_id);
+    let channel_rules = get_channel_rules(env, &user2, community_id, channel_id);
+
+    assert_eq!(
+        community_rules,
+        Rules {
+            enabled: true,
+            accepted: true,
+            text: "No running".to_string(),
+            version: Version::from(1)
+        }
+    );
+
+    assert_eq!(
+        channel_rules,
+        Rules {
+            enabled: true,
+            accepted: true,
+            text: "No jumping".to_string(),
+            version: Version::from(1)
+        }
+    );
+}
+
+fn get_community_rules(env: &mut StateMachine, user: &User, community_id: CommunityId) -> Rules {
+    let summary = client::community::happy_path::summary(env, user, community_id);
+    let selected = client::community::happy_path::selected_initial(env, user, community_id);
+
+    Rules {
+        enabled: summary.rules_enabled,
+        accepted: summary.membership.unwrap().rules_accepted,
+        text: selected.access_rules.text,
+        version: selected.access_rules.version,
+    }
+}
+
+fn get_channel_rules(env: &mut StateMachine, user: &User, community_id: CommunityId, channel_id: ChannelId) -> Rules {
+    let summary = client::community::happy_path::channel_summary(env, user, community_id, channel_id);
+    let selected = client::community::happy_path::selected_channel_initial(env, user, community_id, channel_id);
+
+    Rules {
+        enabled: summary.rules_enabled,
+        accepted: summary.membership.unwrap().rules_accepted,
+        text: selected.access_rules.text,
+        version: selected.access_rules.version,
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq)]
+struct Rules {
+    enabled: bool,
+    accepted: bool,
+    text: String,
+    version: Version,
+}
+
+fn send_dummy_message_with_rules(
+    env: &mut StateMachine,
+    sender: &User,
+    community_id: CommunityId,
+    channel_id: ChannelId,
+    community_rules_accepted: Option<Version>,
+    channel_rules_accepted: Option<Version>,
+) -> community_canister::send_message::Response {
+    client::community::send_message(
+        env,
+        sender.principal,
+        community_id.into(),
+        &community_canister::send_message::Args {
+            channel_id,
+            thread_root_message_index: None,
+            message_id: random_message_id(),
+            content: MessageContentInitial::Text(TextContent { text: "123".to_string() }),
+            sender_name: sender.username(),
+            replies_to: None,
+            mentioned: Vec::new(),
+            forwarding: false,
+            community_rules_accepted,
+            channel_rules_accepted,
+        },
+    )
 }
 
 fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds, controller: Principal) -> TestData {
@@ -116,6 +571,37 @@ fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds, controller
         community_id,
         channel_id,
     }
+}
+
+fn set_community_rules(env: &mut StateMachine, user: &User, community_id: CommunityId, text: String) {
+    let args = community_canister::update_community::Args {
+        name: None,
+        description: None,
+        rules: Some(AccessRules { text, enabled: true }),
+        avatar: OptionUpdate::NoChange,
+        banner: OptionUpdate::NoChange,
+        permissions: None,
+        gate: OptionUpdate::NoChange,
+        public: None,
+        primary_language: None,
+    };
+
+    client::community::happy_path::update_community(env, user, community_id, &args);
+}
+
+fn set_channel_rules(env: &mut StateMachine, user: &User, community_id: CommunityId, channel_id: ChannelId, text: String) {
+    let args = community_canister::update_channel::Args {
+        name: None,
+        description: None,
+        rules: Some(AccessRules { text, enabled: true }),
+        avatar: OptionUpdate::NoChange,
+        permissions: None,
+        gate: OptionUpdate::NoChange,
+        public: None,
+        channel_id,
+    };
+
+    client::community::happy_path::update_channel(env, user, community_id, &args);
 }
 
 #[allow(dead_code)]