+++ backend/canisters/user_index/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Include `use_for_new_canisters` when displaying upgrade proposals ([#4214](https://github.com/open-chat-labs/open-chat/pull/4214))
 - Improve upgrade version check to support multiple active versions ([#4215](https://github.com/open-chat-labs/open-chat/pull/4215))
+- Extend versioned rules to communities and groups ([#4219](https://github.com/open-chat-labs/open-chat/pull/4219))
 
 ## [[2.0.792](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.792-user_index)] - 2023-08-08